# backend/docs/contracts/generated-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/permissions/my-permissions": {
      "get": {
        "operationId": "PermissionsController_getMyPermissions",
        "summary": "Get current user permissions",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Permissions retrieved successfully"
          }
        },
        "tags": ["Permissions"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/permissions/all": {
      "get": {
        "operationId": "PermissionsController_getAllPermissions",
        "summary": "Get all available permissions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Permissions retrieved successfully"
          }
        },
        "tags": ["Permissions"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/permissions/roles": {
      "get": {
        "operationId": "PermissionsController_getAllRoles",
        "summary": "Get all roles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Roles retrieved successfully"
          }
        },
        "tags": ["Permissions"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/permissions/check": {
      "get": {
        "operationId": "PermissionsController_checkPermission",
        "summary": "Check if user has permission",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entity",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "departmentId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "unitId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ownerId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Permission check result"
          }
        },
        "tags": ["Permissions"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/permissions/grant": {
      "post": {
        "operationId": "PermissionsController_grantPermission",
        "summary": "Grant permission to user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Permission granted successfully"
          }
        },
        "tags": ["Permissions"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/permissions/revoke": {
      "post": {
        "operationId": "PermissionsController_revokePermission",
        "summary": "Revoke permission from user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Permission revoked successfully"
          }
        },
        "tags": ["Permissions"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/permissions/roles/assign": {
      "post": {
        "operationId": "PermissionsController_assignRole",
        "summary": "Assign role to user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Role assigned successfully"
          }
        },
        "tags": ["Permissions"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/permissions/roles/{userId}/{roleId}": {
      "delete": {
        "operationId": "PermissionsController_removeRole",
        "summary": "Remove role from user",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role removed successfully"
          }
        },
        "tags": ["Permissions"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/permissions/create": {
      "post": {
        "operationId": "PermissionsController_createPermission",
        "summary": "Create new permission",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Permission created successfully"
          }
        },
        "tags": ["Permissions"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/permissions/roles/create": {
      "post": {
        "operationId": "PermissionsController_createRole",
        "summary": "Create new role",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Role created successfully"
          }
        },
        "tags": ["Permissions"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/permissions/roles/{roleId}/permissions/{permissionId}": {
      "post": {
        "operationId": "PermissionsController_addPermissionToRole",
        "summary": "Add permission to role",
        "parameters": [
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "permissionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Permission added to role successfully"
          }
        },
        "tags": ["Permissions"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/permissions/check": {
      "post": {
        "operationId": "PermissionsController_checkPermission",
        "summary": "Check if user has permission",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Access Control"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/permissions": {
      "post": {
        "operationId": "PermissionsController_createPermission",
        "summary": "Create new permission",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Access Control"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/roles": {
      "get": {
        "operationId": "RolesController_findAll",
        "summary": "List all roles",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Roles"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "RolesController_create",
        "summary": "Create new role",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Roles"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/roles/{id}/permissions/{permissionId}": {
      "delete": {
        "operationId": "RolesController_removePermission",
        "summary": "Remove permission from role",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "permissionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Roles"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/admin/settings": {
      "get": {
        "operationId": "SystemSettingsController_getSettings",
        "summary": "Get all settings",
        "description": "Returns all system settings with optional filtering by category and public visibility. Requires SystemSettings.READ permission.",
        "parameters": [
          {
            "name": "isPublic",
            "required": false,
            "in": "query",
            "description": "Filter for public settings only (true/false)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Filter by setting category",
            "schema": {
              "enum": ["GENERAL", "SECURITY", "NOTIFICATION", "INTEGRATION", "UI"],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Settings retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - SystemSettings.READ permission required"
          }
        },
        "tags": ["System Settings"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "SystemSettingsController_upsertSetting",
        "summary": "Create or update setting",
        "description": "Creates a new setting or updates an existing one. Requires SystemSettings.UPDATE permission and ADMIN or MANAGER role.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "key": {
                    "type": "string",
                    "description": "Setting key (unique identifier)",
                    "example": "site.name"
                  },
                  "category": {
                    "type": "string",
                    "description": "Setting category",
                    "enum": ["GENERAL", "SECURITY", "NOTIFICATION", "INTEGRATION", "UI"]
                  },
                  "value": {
                    "description": "Setting value (any type)",
                    "example": "IRIB Digital Workplace"
                  },
                  "type": {
                    "type": "string",
                    "description": "Data type",
                    "example": "string"
                  },
                  "description": {
                    "description": "Setting description (can be localized)",
                    "example": "{ fa: \"نام سایت\", en: \"Site name\" }"
                  },
                  "isPublic": {
                    "type": "boolean",
                    "description": "Whether setting is publicly accessible",
                    "example": false
                  },
                  "isEditable": {
                    "type": "boolean",
                    "description": "Whether setting can be edited by users",
                    "example": true
                  },
                  "updatedBy": {
                    "type": "string",
                    "description": "User ID who made the change"
                  }
                },
                "required": ["key", "category", "value"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Setting saved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - SystemSettings.UPDATE permission and ADMIN/MANAGER role required"
          }
        },
        "tags": ["System Settings"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "IRIB DWP API",
    "description": "IRIB East Azerbaijan Digital Workplace Platform API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}